Move photo queue model types into camera.model
