+++ asg_client/app/src/test/java/com/mentra/asg_client/camera/model/CurrentRequestTest.java
@@ -1,7 +1,9 @@
-package com.mentra.asg_client.camera;
+package com.mentra.asg_client.camera.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
+
+import com.mentra.asg_client.camera.CameraNeoService;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -14,7 +16,7 @@ public class CurrentRequestTest {
 
     @Test
     public void from_copiesAllFields() {
-        CameraNeo.PhotoCaptureCallback cb = mock(CameraNeo.PhotoCaptureCallback.class);
+        CameraNeoService.PhotoCaptureCallback cb = mock(CameraNeoService.PhotoCaptureCallback.class);
         PhotoRequest pr = new PhotoRequest("/tmp/a.jpg", "medium", true, true, 100_000_000L, cb);
 
         CurrentRequest cur = CurrentRequest.from(pr);
@@ -38,7 +40,7 @@ public class CurrentRequestTest {
     @Test
     public void equals_isReflexive_andCompares_allFields() {
         long t = 42L;
-        CameraNeo.PhotoCaptureCallback cb = mock(CameraNeo.PhotoCaptureCallback.class);
+        CameraNeoService.PhotoCaptureCallback cb = mock(CameraNeoService.PhotoCaptureCallback.class);
         CurrentRequest a = new CurrentRequest("/p", "s", true, 1L, false, t, cb);
         CurrentRequest b = new CurrentRequest("/p", "s", true, 1L, false, t, cb);
         assertThat(a).isEqualTo(b).hasSameHashCodeAs(b);
